# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/arrayelement/samesumelement/{arrLen}/{arr}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "arrayElement"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "arrLen",
                        "name": "arrLen",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "arr",
                        "name": "arr",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/candy/whogetsourcandy/{student}/{candies}/{firstStudent}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Candy"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "student",
                        "name": "student",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "candies",
                        "name": "candies",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "firstStudent",
                        "name": "firstStudent",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/library/bookreturn/{dueDate}/{submitDate}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "dueDate Format DD-MM-YYYY",
                        "name": "dueDate",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "submitDate Format DD-MM-YYYY",
                        "name": "submitDate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}